+++ addon/result_iterator.cc
@@ -2,6 +2,7 @@
 #include "duckdb.hpp"
 #include "duckdb/common/types/hugeint.hpp"
 #include <iostream>
+#include <string.h>
 using namespace std;
 
 namespace NodeDuckDB {
@@ -32,16 +33,16 @@ Napi::Object ResultIterator::Create() { return constructor.New({}); }
 
 typedef uint64_t idx_t;
 
-int32_t GetDate(int64_t timestamp) {
-  return (int32_t)(((int64_t)timestamp) >> 32);
-}
+int64_t GetDate(int64_t timestamp) { return timestamp; }
 
-int32_t GetTime(int64_t timestamp) { return (int32_t)(timestamp & 0xFFFFFFFF); }
+int64_t GetTime(int64_t timestamp) {
+  return (int64_t)(timestamp & 0xFFFFFFFFFFFFFFFF);
+}
 
 #define EPOCH_DATE 719528
 #define SECONDS_PER_DAY (60 * 60 * 24)
 
-int64_t Epoch(int32_t date) {
+int64_t Epoch(int64_t date) {
   return ((int64_t)date - EPOCH_DATE) * SECONDS_PER_DAY;
 }
 
@@ -52,17 +53,25 @@ Napi::Value ResultIterator::FetchRow(const Napi::CallbackInfo &info) {
     return env.Undefined();
   }
   if (!current_chunk || chunk_offset >= current_chunk->size()) {
-    current_chunk = result->Fetch();
+    try {
+      current_chunk = result->Fetch();
+    } catch (const duckdb::InvalidInputException &e) {
+      if (strncmp(e.what(),
+                  "Invalid Input Error: Attempting to fetch from an "
+                  "unsuccessful or closed streaming query result",
+                  50) == 0) {
+        Napi::Error::New(
+            env, "Attempting to fetch from an unsuccessful or closed streaming "
+                 "query result: only "
+                 "one stream can be active on one connection at a time)")
+            .ThrowAsJavaScriptException();
+        return env.Undefined();
+      }
+      throw e;
+    }
     chunk_offset = 0;
   }
-  if (!current_chunk) {
-    Napi::Error::New(
-        env, "No data has been returned (possibly stream has been closed: only "
-             "one stream can be active on one connection at a time)")
-        .ThrowAsJavaScriptException();
-    return env.Undefined();
-  }
-  if (current_chunk->size() == 0) {
+  if (!current_chunk || current_chunk->size() == 0) {
     return env.Null();
   }
   Napi::Value row;
@@ -132,29 +141,32 @@ Napi::Value ResultIterator::getRowObject(Napi::Env env) {
 }
 
 Napi::Value ResultIterator::getCellValue(Napi::Env env, duckdb::idx_t col_idx) {
-  auto &nullmask = duckdb::FlatVector::Nullmask(current_chunk->data[col_idx]);
-  if (nullmask[chunk_offset]) {
+  auto value = current_chunk->data[col_idx].GetValue(chunk_offset);
+  return getMappedValue(env, value);
+}
+
+Napi::Value ResultIterator::getMappedValue(Napi::Env env, duckdb::Value value) {
+  if (value.is_null) {
     return env.Null();
   }
 
-  auto val = current_chunk->data[col_idx].GetValue(chunk_offset);
-  switch (result->types[col_idx].id()) {
+  switch (value.type().id()) {
   case duckdb::LogicalTypeId::BOOLEAN:
-    return Napi::Boolean::New(env, val.GetValue<bool>());
+    return Napi::Boolean::New(env, value.GetValue<bool>());
   case duckdb::LogicalTypeId::TINYINT:
-    return Napi::Number::New(env, val.GetValue<int8_t>());
+    return Napi::Number::New(env, value.GetValue<int8_t>());
   case duckdb::LogicalTypeId::SMALLINT:
-    return Napi::Number::New(env, val.GetValue<int16_t>());
+    return Napi::Number::New(env, value.GetValue<int16_t>());
   case duckdb::LogicalTypeId::INTEGER:
-    return Napi::Number::New(env, val.GetValue<int32_t>());
+    return Napi::Number::New(env, value.GetValue<int32_t>());
   case duckdb::LogicalTypeId::BIGINT:
-    return Napi::BigInt::New(env, val.GetValue<int64_t>());
+    return Napi::BigInt::New(env, value.GetValue<int64_t>());
   case duckdb::LogicalTypeId::HUGEINT: {
     // hugeint_t represents a signed 128 bit integer in two's complement
     // notation napi's BigInt is basically a regular signed integer (MSB) so we
     // want to make sure we pass the absolute value of the huge int into napi
     // plus the sign bit
-    auto huge_int = val.GetValue<duckdb::hugeint_t>();
+    auto huge_int = value.GetValue<duckdb::hugeint_t>();
     int is_negative = huge_int.upper < 0;
     duckdb::hugeint_t positive_huge_int =
         is_negative ? huge_int * duckdb::hugeint_t(-1) : huge_int;
@@ -162,51 +174,71 @@ Napi::Value ResultIterator::getCellValue(Napi::Env env, duckdb::idx_t col_idx) {
     return Napi::BigInt::New(env, is_negative, 2, &arr[0]);
   }
   case duckdb::LogicalTypeId::FLOAT:
-    return Napi::Number::New(env, val.GetValue<float>());
+    return Napi::Number::New(env, value.GetValue<float>());
   case duckdb::LogicalTypeId::DOUBLE:
-    return Napi::Number::New(env, val.GetValue<double>());
+    return Napi::Number::New(env, value.GetValue<double>());
   case duckdb::LogicalTypeId::DECIMAL:
     return Napi::Number::New(
-        env, val.CastAs(duckdb::LogicalType::DOUBLE).GetValue<double>());
+        env, value.CastAs(duckdb::LogicalType::DOUBLE).GetValue<double>());
   case duckdb::LogicalTypeId::VARCHAR:
-    return Napi::String::New(env, val.GetValue<string>());
+    return Napi::String::New(env, value.GetValue<string>());
   case duckdb::LogicalTypeId::BLOB: {
-    int array_length = val.str_value.length();
+    int array_length = value.str_value.length();
     char char_array[array_length + 1];
     // TODO: multiple copies, improve
-    strcpy(char_array, val.str_value.c_str());
+    strcpy(char_array, value.str_value.c_str());
     return Napi::Buffer<char>::Copy(env, char_array, array_length);
   }
   case duckdb::LogicalTypeId::TIMESTAMP: {
-    if (result->types[col_idx].InternalType() != duckdb::PhysicalType::INT64) {
+    if (value.type().InternalType() != duckdb::PhysicalType::INT64) {
       throw runtime_error("expected int64 for timestamp");
     }
-    int64_t tval = val.GetValue<int64_t>();
-    int64_t date = Epoch(GetDate(tval)) * 1000;
-    int32_t time = GetTime(tval);
-    return Napi::Number::New(env, date + time);
-  }
-  case duckdb::LogicalTypeId::DATE: {
-    if (result->types[col_idx].InternalType() != duckdb::PhysicalType::INT32) {
-      throw runtime_error("expected int32 for date");
-    }
-    return Napi::Number::New(env, Epoch(val.GetValue<int32_t>()) * 1000);
+    int64_t tval = value.GetValue<int64_t>();
+    return Napi::Number::New(env, tval / 1000);
   }
   case duckdb::LogicalTypeId::TIME: {
-    if (result->types[col_idx].InternalType() != duckdb::PhysicalType::INT32) {
-      throw runtime_error("expected int32 for time");
+    if (value.type().InternalType() != duckdb::PhysicalType::INT64) {
+      throw runtime_error("expected int64 for time");
     }
-    int64_t tval = val.GetValue<int32_t>();
+    int64_t tval = value.GetValue<int64_t>();
     return Napi::Number::New(env, GetTime(tval));
   }
   case duckdb::LogicalTypeId::INTERVAL: {
-    return Napi::String::New(env, val.ToString());
+    return Napi::String::New(env, value.ToString());
+  }
+  case duckdb::LogicalTypeId::UTINYINT:
+    return Napi::Number::New(env, value.GetValue<uint8_t>());
+  case duckdb::LogicalTypeId::USMALLINT:
+    return Napi::Number::New(env, value.GetValue<uint16_t>());
+  case duckdb::LogicalTypeId::UINTEGER:
+    // GetValue is not supported for uint32_t, so using the wider type
+    return Napi::Number::New(env, value.GetValue<int64_t>());
+  case duckdb::LogicalTypeId::LIST: {
+    auto array = Napi::Array::New(env);
+    for (size_t i = 0; i < value.list_value.size(); i++) {
+      auto &element = value.list_value[i];
+      auto mapped_value = getMappedValue(env, element);
+      array.Set(i, mapped_value);
+    }
+    return array;
+  }
+  case duckdb::LogicalTypeId::STRUCT: {
+    auto object = Napi::Object::New(env);
+    for (size_t i = 0; i < value.struct_value.size(); i++) {
+      auto &child_types = duckdb::StructType::GetChildTypes(value.type());
+      auto &key = child_types[i].first;
+      auto &element = value.struct_value[i];
+      auto child_value = getMappedValue(env, element);
+      object.Set(key, child_value);
+    }
+    return object;
   }
   default:
     // default to getting string representation
-    return Napi::String::New(env, val.ToString());
+    return Napi::String::New(env, value.ToString());
   }
 }
+
 Napi::Value ResultIterator::Close(const Napi::CallbackInfo &info) {
   result.reset();
   return info.Env().Undefined();