+++ CHANGELOG.md
@@ -2,6 +2,77 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.78](https://github.com/deepcrawl/node-duckdb/compare/v0.0.77...v0.0.78) (2021-09-01)
+
+### [0.0.77](https://github.com/deepcrawl/node-duckdb/compare/v0.0.76...v0.0.77) (2021-09-01)
+
+### [0.0.76](https://github.com/deepcrawl/node-duckdb/compare/v0.0.75...v0.0.76) (2021-08-05)
+
+
+### Features
+
+* [ODIN-1051] upgrade duckdb to ver 0.2.8 ([b84117c](https://github.com/deepcrawl/node-duckdb/commit/b84117c3f02e25fd3c7783402bfcec92e57958b4))
+
+
+### Bug Fixes
+
+* cicd bugfix [#114](https://github.com/deepcrawl/node-duckdb/issues/114) ([2a59392](https://github.com/deepcrawl/node-duckdb/commit/2a593924c3730df0bc4757072415af1bb995d369))
+* remove force flag (cicd) [#116](https://github.com/deepcrawl/node-duckdb/issues/116) ([5a30b4f](https://github.com/deepcrawl/node-duckdb/commit/5a30b4feb45963c11d1a8b6b9b74005999273d99))
+
+### [0.0.75](https://github.com/deepcrawl/node-duckdb/compare/v0.0.74...v0.0.75) (2021-07-22)
+
+
+### Bug Fixes
+
+* cicd bugfix [#113](https://github.com/deepcrawl/node-duckdb/issues/113) ([c481f54](https://github.com/deepcrawl/node-duckdb/commit/c481f54bc6be24ec46ea7e147664de46cb655998))
+
+### [0.0.74](https://github.com/deepcrawl/node-duckdb/compare/v0.0.73...v0.0.74) (2021-07-22)
+
+### [0.0.73](https://github.com/deepcrawl/node-duckdb/compare/v0.0.72...v0.0.73) (2021-07-19)
+
+### [0.0.72](https://github.com/deepcrawl/node-duckdb/compare/v0.0.71...v0.0.72) (2021-06-21)
+
+
+### Features
+
+* upgrade to duckdb v. 0.2.7 ([#102](https://github.com/deepcrawl/node-duckdb/issues/102)) ([3832678](https://github.com/deepcrawl/node-duckdb/commit/38326781d75fb125da2a7f417903ac83148c991e))
+
+### [0.0.71](https://github.com/deepcrawl/node-duckdb/compare/v0.0.70...v0.0.71) (2021-06-02)
+
+
+### Features
+
+* [ODIN-895] LIST and STRUCT types ([#99](https://github.com/deepcrawl/node-duckdb/issues/99)) ([6effe12](https://github.com/deepcrawl/node-duckdb/commit/6effe12e997298906bd6c176697b3dd29541cf69))
+
+### [0.0.70](https://github.com/deepcrawl/node-duckdb/compare/v0.0.69...v0.0.70) (2021-06-01)
+
+### [0.0.69](https://github.com/deepcrawl/node-duckdb/compare/v0.0.68...v0.0.69) (2021-05-20)
+
+### [0.0.68](https://github.com/deepcrawl/node-duckdb/compare/v0.0.67...v0.0.68) (2021-05-19)
+
+### [0.0.67](https://github.com/deepcrawl/node-duckdb/compare/v0.0.66...v0.0.67) (2021-04-07)
+
+
+### Bug Fixes
+
+* [ODIN-512] bumping logic ([#91](https://github.com/deepcrawl/node-duckdb/issues/91)) ([27b412d](https://github.com/deepcrawl/node-duckdb/commit/27b412d1967834f549261443bfd474a8d49297af))
+
+### [0.0.66](https://github.com/deepcrawl/node-duckdb/compare/v0.0.64...v0.0.66) (2021-04-06)
+
+### [0.0.64](https://github.com/deepcrawl/node-duckdb/compare/v0.0.61...v0.0.64) (2021-03-26)
+
+
+### Features
+
+* duckdb v  6fe573 ([#74](https://github.com/deepcrawl/node-duckdb/issues/74)) ([3b5ceca](https://github.com/deepcrawl/node-duckdb/commit/3b5cecad3fc5b692ca13c35cb4b76d5440604810))
+
+### [0.0.61](https://github.com/deepcrawl/node-duckdb/compare/v0.0.59...v0.0.61) (2021-02-03)
+
+
+### Features
+
+* duckdb v  6fe573 ([#74](https://github.com/deepcrawl/node-duckdb/issues/74)) ([3b5ceca](https://github.com/deepcrawl/node-duckdb/commit/3b5cecad3fc5b692ca13c35cb4b76d5440604810))
+
 ### [0.0.59](https://github.com/deepcrawl/node-duckdb/compare/v0.0.58...v0.0.59) (2020-12-16)
 
 ### [0.0.58](https://github.com/deepcrawl/node-duckdb/compare/v0.0.57...v0.0.58) (2020-12-16)